Dynamic model resolution test just logic failing
Need to fix all the failing tests and make sure the coverage is up to
snuff. When the tests are okay this should be fine to be used.

Probably will want to do some simple admin views that can be copied in
and then proper documentation together with getting some help with
setting proper names.